Implement a pre-processor for entities and infer data type IDs
