+++ apps/hash-graph/libs/api/src/rest/data_type.rs
@@ -37,7 +37,6 @@ use graph::{
     subgraph::identifier::DataTypeVertexId,
 };
 use graph_types::{
-    knowledge::ValueWithMetadata,
     ontology::{
         DataTypeId, DataTypeMetadata, DataTypeWithMetadata, OntologyTemporalMetadata,
         OntologyTypeClassificationMetadata, OntologyTypeMetadata, OntologyTypeReference,
@@ -110,8 +109,6 @@ use crate::rest::{
             Conversions,
             Operator,
             Variable,
-
-            ValueWithMetadata,
         )
     ),
     tags(
