+++ apps/hash-graph/libs/api/src/rest/entity.rs
@@ -44,10 +44,14 @@ use graph_types::{
             SourceProvenance, SourceType,
         },
         link::LinkData,
-        ArrayMetadata, Confidence, EntityTypeIdDiff, ObjectMetadata, Property, PropertyDiff,
-        PropertyMetadata, PropertyMetadataObject, PropertyObject, PropertyPatchOperation,
-        PropertyPath, PropertyPathElement, PropertyProvenance, PropertyWithMetadata,
-        PropertyWithMetadataObject, ValueMetadata,
+        property::{
+            ArrayMetadata, ObjectMetadata, Property, PropertyDiff, PropertyMetadata,
+            PropertyMetadataObject, PropertyObject, PropertyPatchOperation, PropertyPath,
+            PropertyPathElement, PropertyProvenance, PropertyWithMetadata,
+            PropertyWithMetadataArray, PropertyWithMetadataObject, PropertyWithMetadataValue,
+            ValueMetadata,
+        },
+        Confidence, EntityTypeIdDiff,
     },
     owned_by_id::OwnedById,
     Embedding,
@@ -88,6 +92,8 @@ use crate::rest::{
         schemas(
             CreateEntityRequest,
             PropertyWithMetadata,
+            PropertyWithMetadataValue,
+            PropertyWithMetadataArray,
             PropertyWithMetadataObject,
             ValidateEntityParams,
             CountEntitiesParams,
