+++ libs/@local/hash-graph-types/rust/src/knowledge/link.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::knowledge::{entity::EntityId, Confidence, PropertyProvenance};
+use crate::knowledge::{entity::EntityId, property::PropertyProvenance, Confidence};
 
 /// The associated information for 'Link' entities
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
