+++ libs/@local/hash-graph-types/rust/src/knowledge/property/metadata/value.rs
@@ -1,9 +1,9 @@
 use serde::{Deserialize, Serialize};
 use type_system::url::VersionedUrl;
 
-use crate::knowledge::{Confidence, PropertyProvenance};
+use crate::knowledge::{property::PropertyProvenance, Confidence};
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Default, Clone, PartialEq, Serialize, Deserialize)]
 #[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
 pub struct ValueMetadata {
@@ -20,7 +20,7 @@ pub struct ValueMetadata {
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct ValueWithMetadata {
+pub struct PropertyWithMetadataValue {
     pub value: serde_json::Value,
     pub metadata: ValueMetadata,
 }
