+++ tests/hash-graph-benches/read_scaling/knowledge/complete/entity.rs
@@ -25,7 +25,7 @@ use graph_types::{
     knowledge::{
         entity::{Entity, ProvidedEntityEditionProvenance},
         link::LinkData,
-        PropertyObject, PropertyProvenance, PropertyWithMetadataObject,
+        property::{PropertyObject, PropertyProvenance, PropertyWithMetadataObject},
     },
     owned_by_id::OwnedById,
 };
