+++ tests/hash-graph-benches/representative_read/seed.rs
@@ -13,7 +13,7 @@ use graph_types::{
     knowledge::{
         entity::{EntityUuid, ProvidedEntityEditionProvenance},
         link::LinkData,
-        PropertyObject, PropertyProvenance, PropertyWithMetadataObject,
+        property::{PropertyObject, PropertyProvenance, PropertyWithMetadataObject},
     },
     owned_by_id::OwnedById,
 };
