+++ tests/hash-graph-integration/postgres/drafts.rs
@@ -9,8 +9,10 @@ use graph_test_data::{data_type, entity, entity_type, property_type};
 use graph_types::{
     knowledge::{
         entity::{EntityId, ProvidedEntityEditionProvenance},
-        Property, PropertyObject, PropertyPatchOperation, PropertyPath, PropertyWithMetadata,
-        PropertyWithMetadataObject,
+        property::{
+            Property, PropertyObject, PropertyPatchOperation, PropertyPath, PropertyWithMetadata,
+            PropertyWithMetadataObject,
+        },
     },
     owned_by_id::OwnedById,
 };
